backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23

openai>=1.30.0

faiss-cpu==1.8.0.post1

pypdf==4.3.1
pdfplumber==0.11.4
PyMuPDF==1.24.10

python-multipart==0.0.9

numpy==1.26.4
tqdm==4.66.5
python-dotenv==1.0.1

tenacity>=8.2.3
requests>=2.31.0